separate function send and reply message
